requirements/test.txt
#
# This file is autogenerated by pip-compile with Python 3.8
# by the following command:
#
#    make upgrade
#
asgiref==3.6.0
    # via
    #   -r requirements/base.txt
    #   django
async-timeout==4.0.2
    # via
    #   -r requirements/base.txt
    #   redis
attrs==22.2.0
    # via
    #   -r requirements/base.txt
    #   openedx-events
    #   pytest
certifi==2022.12.7
    # via
    #   -r requirements/base.txt
    #   requests
cffi==1.15.1
    # via
    #   -r requirements/base.txt
    #   cryptography
    #   pynacl
charset-normalizer==3.1.0
    # via
    #   -r requirements/base.txt
    #   requests
click==8.1.3
    # via
    #   -r requirements/base.txt
    #   code-annotations
    #   edx-django-utils
code-annotations==1.3.0
    # via -r requirements/test.in
coverage[toml]==7.2.2
    # via pytest-cov
cryptography==39.0.2
    # via
    #   -r requirements/base.txt
    #   pyjwt
ddt==1.6.0
    # via -r requirements/test.in
    # via
    #   -c https://raw.githubusercontent.com/edx/edx-lint/master/edx_lint/files/common_constraints.txt
    #   -r requirements/base.txt
    #   django-crum
    #   django-extensions
    #   django-filter
    #   django-model-utils
    #   djangorestframework
    #   drf-jwt
    #   edx-django-release-util
    #   edx-django-utils
    #   edx-drf-extensions
    #   edx-rbac
    #   jsonfield2
    #   openedx-events
django-crum==0.7.9
    # via
    #   -r requirements/base.txt
    #   edx-django-utils
    #   edx-rbac
django-extensions==3.2.1
    # via -r requirements/base.txt
django-filter==22.1
    # via -r requirements/base.txt
django-model-utils==4.3.1
    # via
    #   -r requirements/base.txt
    #   edx-rbac
django-simple-history==3.0.0
    # via
    #   -c https://raw.githubusercontent.com/edx/edx-lint/master/edx_lint/files/common_constraints.txt
    #   -r requirements/base.txt
django-waffle==3.0.0
    # via
    #   -r requirements/base.txt
    #   edx-django-utils
    #   edx-drf-extensions
djangoql==0.17.1
    # via -r requirements/base.txt
djangorestframework==3.14.0
    # via
    #   -r requirements/base.txt
    #   drf-jwt
    #   edx-drf-extensions
drf-jwt==1.19.2
    # via
    #   -r requirements/base.txt
    #   edx-drf-extensions
edx-django-release-util==1.2.0
    # via -r requirements/base.txt
edx-django-utils==5.3.0
    # via
    #   -r requirements/base.txt
    #   edx-drf-extensions
edx-drf-extensions==8.4.1
    # via
    #   -r requirements/base.txt
    #   edx-rbac
edx-opaque-keys[django]==2.3.0
    # via
    #   -r requirements/base.txt
    #   edx-drf-extensions
    #   openedx-events
edx-rbac==1.7.0
    # via -r requirements/base.txt
exceptiongroup==1.1.1
    # via pytest
factory-boy==3.2.1
    # via -r requirements/test.in
faker==18.2.0
    # via factory-boy
fastavro==1.7.3
    # via
    #   -r requirements/base.txt
    #   openedx-events
future==0.18.3
    # via
    #   -r requirements/base.txt
    #   pyjwkest
idna==3.4
    # via
    #   -r requirements/base.txt
    #   requests
iniconfig==2.0.0
    # via pytest
jinja2==3.1.2
    # via code-annotations
jsonfield2==4.0.0.post0
    # via -r requirements/base.txt
markupsafe==2.1.2
    # via jinja2
mysqlclient==2.1.1
    # via -r requirements/base.txt
newrelic==8.7.1
    # via
    #   -r requirements/base.txt
    #   edx-django-utils
openedx-events==7.0.0
    # via -r requirements/base.txt
packaging==23.0
    # via pytest
pbr==5.11.1
    # via
    #   -r requirements/base.txt
    #   stevedore
pluggy==1.0.0
    # via pytest
ply==3.11
    # via
    #   -r requirements/base.txt
    #   djangoql
psutil==5.9.4
    # via
    #   -r requirements/base.txt
    #   edx-django-utils
pycparser==2.21
    # via
    #   -r requirements/base.txt
    #   cffi
pycryptodomex==3.17
    # via
    #   -r requirements/base.txt
    #   pyjwkest
pyjwkest==1.4.2
    # via
    #   -r requirements/base.txt
    #   edx-drf-extensions
pyjwt[crypto]==2.6.0
    # via
    #   -r requirements/base.txt
    #   drf-jwt
    #   edx-drf-extensions
pymongo==3.13.0
    # via
    #   -r requirements/base.txt
    #   edx-opaque-keys
pynacl==1.5.0
    # via
    #   -r requirements/base.txt
    #   edx-django-utils
pytest==7.2.2
    # via
    #   pytest-cov
    #   pytest-django
pytest-cov==4.0.0
    # via -r requirements/test.in
pytest-django==4.5.2
    # via -r requirements/test.in
python-dateutil==2.8.2
    # via
    #   -r requirements/base.txt
    #   edx-drf-extensions
    #   faker
python-slugify==8.0.1
    # via code-annotations
pytz==2022.7.1
    # via
    #   -r requirements/base.txt
    #   django
    #   djangorestframework
pyyaml==6.0
    # via
    #   -r requirements/base.txt
    #   code-annotations
    #   edx-django-release-util
redis==4.5.2
    # via -r requirements/base.txt
requests==2.28.2
    # via
    #   -r requirements/base.txt
    #   edx-drf-extensions
    #   pyjwkest
rules==3.3
    # via -r requirements/base.txt
semantic-version==2.10.0
    # via
    #   -r requirements/base.txt
    #   edx-drf-extensions
six==1.16.0
    # via
    #   -r requirements/base.txt
    #   edx-django-release-util
    #   edx-drf-extensions
    #   edx-rbac
    #   pyjwkest
    #   python-dateutil
sqlparse==0.4.3
    # via
    #   -r requirements/base.txt
    #   django
stevedore==5.0.0
    # via
    #   -r requirements/base.txt
    #   code-annotations
    #   edx-django-utils
    #   edx-opaque-keys
text-unidecode==1.3
    # via python-slugify
tomli==2.0.1
    # via
    #   coverage
    #   pytest
urllib3==1.26.15
    # via
    #   -r requirements/base.txt
    #   requests
